+++ packages/api-client/src/main/Client.ts
@@ -17,8 +17,8 @@
  *
  */
 
-const logdown = require('logdown');
 import {AxiosResponse} from 'axios';
+import * as logdown from 'logdown';
 import {AssetAPI} from './asset/';
 import {AccessTokenData, AuthAPI, Context, LoginData, RegisterData} from './auth';
 import {AccessTokenStore} from './auth/';
@@ -41,12 +41,12 @@ import {UserAPI} from './user/';
 const {version}: {version: string} = require('../../package.json');
 
 class APIClient {
-  private readonly logger: any = logdown('@wireapp/api-client/Client', {
+  private readonly logger = logdown('@wireapp/api-client/Client', {
     logger: console,
     markdown: false,
   });
 
-  private readonly STORE_NAME_PREFIX: string = 'wire';
+  private readonly STORE_NAME_PREFIX = 'wire';
   // APIs
   public asset: {api: AssetAPI};
   public auth: {api: AuthAPI};